+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 1.7.28 (build 21409, api 8, 2023-10-03)
+### 1.7.28 (build 21410, api 8, 2023-10-03)
 
 - Massively cleaned up code related to rendering and window systems (OpenGL,
   SDL, etc). This code had been growing into a nasty tangle for 15 years
@@ -95,7 +95,8 @@
   Thanks SoK05 and Dliwk!
 - In cases where there's no browser available, the v2 account sign-in URL can
   now be tapped to copy it (Thanks vishal332008!).
-  
+- Removed the bits from `babase.app` that were deprecated in 1.7.27. I know that
+  was only one version ago, but this version has been cooking for a while now.
   
 ### 1.7.27 (build 21282, api 8, 2023-08-30)